# Optimize Arm int8 GEMM using UDOT-by-element