--- a/src/gemm/kernels/wasm.rs
+++ b/src/gemm/kernels/wasm.rs
@@ -290,7 +290,7 @@ unsafe impl Kernel<u8, i8, i32> for WasmInt8Kernel {
         let (b, b_col_sums) = packing::int8::extract_packed_b::<{ Self::NR }>(b);
 
         const NR_REGS: usize = WasmInt8Kernel::NR / X32_LANES;
-        simd_int8_gemm::<_, { Self::MR }, { Self::NR }, NR_REGS>(
+        simd_int8_gemm::<_, _, { Self::MR }, { Self::NR }, NR_REGS>(
             self.isa,
             tile_ptr,
             tile_row_stride,