@robertknight 2025-04-28
Tested with ModernBERT Base on an M3, approximately a 9% improvement.

```
cargo run -p rten-cli -r -- modernbert-base.quant.rten -s sequence_length=512 -n 10 -p
```

Before: ~302ms/iter
After: ~275ms/iter

Or using the GEMM benchmark:

```
 cargo test -p rten --release -- --ignored --nocapture bench_gemm_mix
```

The M=N=K=1024 case goes from 1.1 TFLOPS -> 1.4 TFLOPS.